requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
pymongo==4.6.1
PyJWT==2.8.0
Werkzeug==3.0.1
reportlab==4.0.7
python-dotenv==1.0.0
